Add v1 support for taskrun and pipelinerun cancel
This patch adds v1 support for taskrun and pipelinerun
cancel command and also does few refactoring

Part of: #1804
